Incomplete attempt to make QueryReifier inherit from Exp[Traversable]
QueryReifier extends Exp[QueryReifier] is odd. Making it a child of
Exp[Traversable] instead means getting Exp[Traversable] as result of other
methods, so that QueryReifier method are not available without casting.
To solve this, I move those methods to Exp, and make them conditional on T being
Traversable[T']. To define such 'conditional method', I encode a typeclass
constraint on T using implicits, like for Exp.+ and so on.

However, when doing so, type inference cannot any more figure out the type of
anonymous functions passed to map, flatMap and withFilter, and this problem
cannot be solved in this encoding (see comment on Exp.map). That's why I stop
developing this approach; however, this approach would still make sense if the expression
trees were built by a compiler plugin - so that type inference would happen
before conversion of ASTs to expression trees.

The code is not complete just because I need to update the declarations of
exec() to their new type.