OpenEnconding: design comments and debugging for toExpTempl
Some important comments derive from "Fighting Bit Rot with Types (Experience
Report: Scala Collections)", by Odersky and Mooors. CanBuildExp (like
CanBuildFrom) is chosen
according to the static type of the original expression, thus should only determine the
static type of the result, but it should delegate to the original expression to
ensure that the dynamic type of the result depends on the dynamic type of the
source expression.
Moreover, add a much reduced test case showing the problem with toExpTempl, and
more experiments. I'm still not entirely comfortable posting this someplace
public, though.