Allow chaining withFilter invocation - with tests
This was not possible for a previous oversight - we need an automatic
conversion for values, and FilterMonadicOps needed to be adjusted
to allow that. Interestingly, this shows that the first version
I wrote of FilterMonadicOps, having
t: Exp[FilterMonadic[T, Repr]], was correct in one case, and the
second (t: Exp[Repr]) was correct in the other. Thus I have
added a further type parameter to FilterMonadicOpsLike, a bit like
TraversableLikeOps.