SimpleOpenEncoding: show that the encoding is actually open
SimpleOpenEncodingBase imports all modules and defines the
conversion, therefore can be written only using a closed-world
assumption.
However, that's in no way necessary: we can move the list of
used modules to the actual client - which needs to know only
which modules it uses.
For this, we also move implicit conversions in their specific
modules.

Interesting Scala trivia: when conversion for Map[K, V] are in
a _superclass_ of the class defining conversions for Traversable[T],
the Scala compiler complains they are ambiguous. The ones for
Traversable have higher priority because they are defined in a
subclass, but the other ones have a more specific type. However,
it seems that this situation is not realistic for us.
Moreover, I'm sure we can simply use objects and imports rather
than traits and inheritance here - I'll show this in next commit.