Fix the return type of force
The problem was that TraversableViewOps[T, Repr] was
<: TraversableViewLikeOps[T, Traversable[T], TraversableView[T, Repr]]
rather than
<: TraversableViewLikeOps[T, Repr, TraversableView[T, Repr]]
The problem was reflected in force returning Exp[Traversable[T]]
instead of Exp[Repr].
The second type could not be used because of a type bound
(as explained in the source), which however I discovered
to be now superfluous. Hence I removed it and solved this problem.