Stop useless stringification for Var names
Profiling showed that "v" + varCounter during FuncExp creation
was one of the biggest execution hotspots - ~30% of the overall
benchmark execution, and thus an even bigger share of the
interpretation overhead. Why are FuncExp created in inner loops?
That happens e.g. here:
 coll.flatMap(x => x.methods.map(f))
For each x in coll, the execution of map will create FuncExp(f).