Help type inference, way one: type synonyms are a problem
Type inference here has to unify Exp[Traversable[(Method_Info, Instruction)]]
with Exp[Traversable[SND[Instruction]]] - but it would need to invent
SND just because it's in scope, which is a bit too much; it would be easier
with a "newtype".
Yet, with a type annotation, this works! I can comment out both the type
parameters to expToTypeFilterOps and its call itself!