BasicTests: add Exp.is(Null) method to avoid writing Const in the user program
Solve the problem of last commit.

* Is this the best approach to the solution? I cannot think of anything best.

The alternative would be to check within the implementation of Eq,
but I'm not convinced it's better, and adds an conditional jump in the
implementation of equality, so I feel it might be slower.

* What if we have a null value at runtime?
The existing overload of Exp.is will be invoked, the argument type will not
be Null so the typechecker will insert a call to toExp and everything works.

* Should we cache a Const(null) instance, to save memory?
Maybe.
* Should we ensure that only one Const(null) instance can exist, to ensure that
toExp(null) eq toExp(null), like null eq null?
No: that's expensive and we don't have any support for AnyRef.eq on
expressions anyway.
* Do we need to overload other methods of our DSEL as well?
In theory yes, but which other methods return normally when receiving null?
I can think of none, except != which we don't support directly.