Update code to use caching infrastructure
UpdateVal cannot describe the event where no old value was available.
Add the message NewVal for that case.

XXX: it is not clear whether that case should be allowed: should we do IVM
on nodes which might not have been initialized? It sounds sensible on
Mynot, but not elsewhere, so maybe it's wrong also on Mynot.