Remove structure micro-optimization
OpsExpressionTree is just an object extending a trait.
Stop using it - the only potential reason is a microoptimization,
since imports have a runtime impact when modules are expressed as
traits. However, the JIT compiler might optimize that away, so only
benchmarking might show whether an impact is there or not.

OTOH, trait inheritance might be overkill for the structuring task at
hand. We use traits as modules only to be able to re-export imported
symbols easily.