Disable IVM for not() and exists()
I'm also going to disable IVM in more places to improve performance.
I also remove the not() method, which was also used to avoid an explicit
dependency on the Not case class. Instead, now I use ! which is
imported from BaseTypesOps - this is now possible after that
Lifting was split up a bit.