Commutative operators: reduce memory footprint
It looks nice but it is expensive to pass op to the constructor.
Rather make it a method - it's almost as compact but consumes no memory
in instances, thanks to type inference.
Note that if op is made a val instead of a def, type inference
doesn't work (even with override).