Fix test failure
Implicit parameters of case classes are not compared for equality; for
IfInstanceOf and TypeFilter, this is a significant problem.

Thus, make the value derived from the implicit parameter a
regular case class parameter, and change all clients to accomodate that
change. Note that the user code is not changed, only the internal code
which deals with IfInstanceOf/TypeFilter.

Reflection 1: with case classes, it's easier to alter a bit the client
than to explicitly generate the code.
Reflection 2: notice how type inference now fails without a good reason
in TraversableOps, possibly because now TypeFilter.apply is overloaded.