PartialFuncExp: consolidate code
interpret(), in essence, duplicated Function.unlift and
FuncExp.interpret() - let's avoid that.

Moreover, we've seen that FuncExp.interpret() now doesn't work
any more, because it uses substitution, which does not work
together with the environment-based evaluation used by default
(in FuncExpInt).
I assume this code duplicated also that bug (which is not
trivial to reproduce), so this consolidation should also
fix it.