Implicitly convert Seq[Exp[T]] => Exp[Seq[T]], use this for indexes instead of tuples
This conversion is useful whenever using Seq(a, b, c) where a, b, c are
Exp's.
An expression in Tutorial needs to be altered and gets a better type,
but that was unexpected: this happens because an invocation of
Seq[A].map(f: A => Exp[B]): Seq[Exp[B]] is in a context where Exp[_] is
expected, and now, instead of Const, a more specific conversion (the new one)
is used.

(An unrelated change improved the type of the result from Traversable to Seq).