Test indexing on FindBugsAnalyses.PublicFinalizer
The existing query uses exists, which is not supported by indexing;
we can easily avoid that to get a different query.
We define an index, and the optimizer recognizes that the defined index
is applicable, but execution fails for a bug.

Namely, we lookup on index: Map[K, V] using index.apply(key): V
instead of index.get(key): Option[V], hence whenever
!index.contains(key) (as in this case) we get an exception.
This should be easily fixable.

Moreover, remember that we use Seq and not tuples for the index idx.
Hence we invoke groupBy on an expression of type Exp[Seq[Seq[T]]],
where T = lub(ClassFile, Method) = AnyRef with ...
Hence, the argument of groupBy when defining idx needs to use a cast.