Fix bug introduced in refactoring
In tuplingTransform, before commit 6336a7839dff70d38d51d3ad975f79c55a9a55d9,
v was bound to an argument and used. In that commit, I shadowed that variable
through pattern matching, and thus the internal reference referred to another
variable. This was discovered only thanks to tests in SubquerySharingTests
(which is confusing, given that indexing is also used in FindBugsAnalysis).