Reimplement Let and Some without onExp
We already introduced the conversion from Option[Exp[T]] to
Exp[Option[T]]; now we remove the old implementation of Let and the source
code related to it.

We might set in next commit Let = Seq rather than Let = Some; some
client code is changed to allow for that, by using Some instead of Let
where required. Sometimes I use asExp(Some(x)), but never in queries
the user needs to write, only as expected results of transformation.

Some test fails because ExpOption is buggy; it will be fixed in a
separate commit.