Write specific test for the previous failure
Writing this test was hard, because the bug breaks substitution and also
impairs equality comparison: the body of ExpOption is not subject to
substitution, hence vars are not alpha-renamed during transformation in the
ExpOption body.
This bug is hidden because variables are also not alpha-renamed during
equality comparison - hence v1 => ExpOption(Some(v1)) seems the same as
v2 => ExpOption(Some(v1)).