IndeAllow calling TypeMapping.get with Class parameter
For indexing, we need to call get on a type index starting from a Class,
not a ClassManifest (since the former is what a TypeCase provides).

Also update the lifting of TypeMapping.get.

Note that this does not pass around MaybeSub - that's not yet in the
lifted interface at all. This will be fixed.