Fix 'phantom' type annotation - tupleVar had the wrong type
tupleVar: TypedVar[Seq[U]] is an artifact of when we did not use tuples
but only sequences. Since the error is just in an erased type argument,
and both the producer and the consumer of the value with the wrong
type are generic and erased, this error will only cause 'heap pollution'
but no actual runtime error (in fact, this error also happens in
well-tested indexing code from SubquerySharing).