Fix bug introduced in commit 14b20be50713e695a60fb2f9231f6103332eb48c
That commit broke optimiations, this is the fix.

I changed ExpSeq from
  case class ExpSeq[T](children: Exp[T]*)
to:
  case class ExpSeq[T](children: Seq[Exp[T]])
and added a compatibility constructor, but I forgot to update pattern
matches! Now a match with "case ExpSeq(x) =>" binds x to be
a list of arbitrary length, instead verifying that ExpSeq.children contains
only one element and binding x to it.