Try adding type safety to problematic optimizations
Here I use the standard trick of writing a parametric function containing
the body of the optimization.

This doesn't prevent the previous problem, unfortunately.
I can still forget, in flatMapToMap, to write Seq in the pattern match
and get the code to compile - probably because U is deduced to be Any
and body, of type Seq[Exp[_]], is implicitly converted to Exp.

Annotating body with Exp[whatever] adds a type pattern which simply
fails here - and how should the compiler know that Exp[_] with Seq[_]
is not inhabited, without forbidden closed-world reasoning?
In fact, IVM builds collections which are also expressions, hence
it makes even sense for Exp[_] with Seq[_] to be inhabited.
And of course Scalac should not complain that I try to match a Seq
against Exp[_] - that's the point of type patterns, usually!