Make typeCase return a Set, to be coherent with its implementation
The implementation can change the return order, hence it's a set.
Actually, we could live with a bag, but we have no representation
for those! So let's just use a set.
Additionally, we need to change some queries: a comprehension
with a generator ranging over a commutative collection (e.g. a set)
must be commutative.