Remove mention of non-original GPL code
The code in project/Generator.scala is a trivial client of Scalate.
Hence I claim that the original code fragment is derivative work
of other code, but not original enough to warrant separate copyright.

See discussion in http://en.wikipedia.org/wiki/Derivative_work:
"For copyright protection to attach to a later, allegedly derivative work,
it must display some originality of its own."