Workaround incremental recompilation bug
When package object squopt is not commented out and is recompiled
incrementally because some dependency changed, I get very often (always?)
the error:

```
[info] Compiling 1 Scala source to /Users/pgiarrusso/Documents/Research/Sorgenti/linqonsteroids/target/scala-2.9.2/classes...
[error] type _$24 is defined twice
[error] one error found
```

without extra information.

To avoid this, avoid using a package object, at least during development.
This could even be reverted for releases to match the documentation -
if wanted.