Try fixing failed test
The introduction of constant folding, it seems, made this test fail.
So, first, alter the test.
A further failure arises because constant folding generates the wrong
kind of node - Const instead of ConstByIdentity.

The code used to fix this doesn't typecheck without the cast
`.asInstanceOf[Exp[T]]`, which is weird.

However, I'll have to revert this fix: using ConstByIdentity here
makes the index lookup fail, since it reduces the available equalities.
Not sure what's the better fix in general.