Optimizations: Move away dead code/experiments to preserve
Many of these experiments are relevant to research - for instance, they
show that GADT support for transformation is far from perfect. So they
should be kept. But not together with other transformations.

Hence, move most of the dead code to another file, delete some of it,
rewrite some other comments to clarify them, & some related general
cleanup.