Fix indexing test-case breakage
The last broken test-cases I had were about indexing failing.
The problem was that the subquery sharing algorithms tried to look up
indexes which were not in the same normal form as the indexes in the
index repository. In particular, the indexes looked up were not
_any more_ in normal form after I introduced new optimization rules
which were applied to indexes before storing them in the index
repository.

A robust solution for this would be to prove that the optimizations
produce a normal form, and then just normalize both indexes before
storing them in the repo and indexes before looking them up in the repo.
Also, that would make for cool theory.