Don't call head on potentially empty lists
The code called head on a potentially empty list. Using headOption
is in this case a simple and elegant solution.
This bug showed up after changing the optimizer - the affected
optimizations were called, for no apparent reason, after flatMapToMap,
even though they match flatMap nodes. That might have been on purpose
to avoid this problem, but was not explained and seems too fragile.