Function lifting: generate code and have split IDs
Now, the ID is split into a method name and the rest of the ID.
Only the method name is output when generating code.
The rest of the ID is kept optional for now.
To ease the transition, I now generate the overloads of fmap - first
I tried fixing the definitions by hand, but scalac's impredictable
constraints make it too hard.