Exp.children: convert from Seq to List, to use mapConserve in transform
Seq is List by default anyway, so the scope of the change is smaller
than it seems.
But since transform is used so much in the optimizer, it makes sense
to consider not rebuilding nodes.
So in transform we use not only mapConserve, but also try to skip
calling genericConstructor.