Invert test logic
The test in the SQuOpt queries is ==#, while it should be !=# like in
native queries. Of course, this only showed up on the JDK, of all
testcases I have.

After a couple hours of debugging (thanks to the slowness of IDEA), I
figured it out. And I damn _had_ compared the query source code before,
multiple
This was introduced in commit 5e06cd8a246c16c2ea8ee11fc24395ae6d777270.