Switch to Scala 2.10.0-M6 (doesn't compile)
The error message you will see is a compiler regression, not yet fixed in
compiler snapshots. A workaround follows in next commit.

The switch is needed to incorporate (hopefully) the fix to this issue, which we
are now facing:
https://issues.scala-lang.org/browse/SI-6086

We will have to be careful to Vector, since -M6 has a big performance regression
for it.