Rename Exp.find to avoid conflicts with Matcher.find()
Having to do such renamings is annoying: methods in Exp share the same namespace
as all methods in all wrapped classes. I hope that using protection also works,
that is, that if methods on Exp are not visible at the caller site they do not
cause conflicts. However I'll try that at some later point.