Fix NullPointerException due to option parsing
Currently FBAnalysesBase uses zipFiles before it is set by the
constructor of its subclass, and this can't be changed easily (without
introducing lazy vals, I'm worried about their performance impact).

So, set zipFiles must be set _before_ calling the constructor - do the
option parsing in the caller. This is somewhat unsatisfying because the
option names and types are duplicated, but at least their defaults are
not.