Proper fix to problems with macros
This change turns macroId into a real identity macro:
-  def macroId(arg: Any) = macro macroId_impl
-  def macroId_impl(c: Context)(arg: c.Expr[Any]): c.Expr[Any] = arg
+  def macroId[T](arg: T): T = macro macroId_impl[T]
+  def macroId_impl[T: c.AbsTypeTag](c: Context)(arg: c.Expr[T]): c.Expr[T] = arg
Apply the same change wherever appropriate.
The problem is in fact orthogonal to macros. `def idAny(v: Any): Any =
v` is similarly problematic, because in `idAny expr`, `expr` is
typechecked against expected type `Any`, not against an unknown expected
type; in some cases, this makes a difference, because the typechecker
will try to give to make the expression have type `Any`, instead of
allowing a subtype.

* Additionally, some debug output is made more readable.