Start adapting code to rearchitecturing
This commits essentially replaces Exp by Def whenever appropriate,
and starts inserting calls to Sym in the few pattern matches left in.

Moreover, I needed to fix some return types in tupleSupport - they declared to
return not Exp but a concrete subclass (which was stupid before and broken now).

In some cases (in tupleSupport), I need to add more type parameters to help type
inference (the alternative would be to call an implicit conversion by name).

The code compiles with a set of other hacky changes which I'll commit
next.