Bugfix: Make FuncExpIdentity work again
With CSE, f.body is an Exp[_], but f.x is a Def[_], so comparing
them should always fail.
Confusingly, fixing this even fixes an optimization bug in an
evaluation query; I expect that's the case because it fixes some
optimizations which prevent other buggy ones from applying (not sure
which ones).