Fix test for SBT (and break it inside IDEA)
Conceptually, the test passes in both environments, producing
slightly different code in the two cases. Since the code links with
the compiler and IDEA has a special test runner (which might link the
compiler as well), after all the difference in result is not too
surprising. Still, of course, it's annoying.
But I have no better solution for now.