Hacky fix to GroupByTests
The type annotation in the changed line causes a failure in compilation!
In the output code, the underlying collection has static type Vector[Int], so
indexBy [Repr = Vector[Int]] requires an argument of type
CanBuildFrom[Vector[Int], Int, Vector[Int]].

In the compiled code the argument however does not satisfy that type, because
the CanBuildFrom instance used by indexBy has type
CanBuildFrom[Traversable[Int], Int, Traversable[Int]]. This is not a subtype of
the expected type because CanBuildFrom's 3rd parameter is covariant while
Traversable[Int] is a supertype of Vector[Int].

The underlying problem is that indexBy should allow for That to be different
from Repr; this is fixed in next commit.

I wonder how to express the property that is violated by the current interface.
Probably, method types should be covariant in their arguments: after staging,
the static type of an expression can be more specific, but invocations on it should
still be valid even if the types of other parameters are not correspondingly
refined. The current interface for indexBy does not allow that, while the
interface in next commit does.