Refactoring: move inner functions out of toValueCSE
This is needed so that in a future commit, I'll be able to call these
functions only from precompiledExpToValue, when lookup in expCodeCache
failed and it must be compiled. This would remove the CSE overhead
in the fastpath.
To keep the current behavior, toValueCSE clear the definitions map.

Note: View this patch with git diff -w.