Replace Exp[_] with Exp[Any] in many places
This is the first batch of such changes I'm testing, it affects most of
the transformation infrastructure.

Rationale: Scalac has poor support for Exp[_]. Exp[Any] is better
because Exp is covariant. We also have functions from Exp[_], but they
are mostly Exp transformers, which are in fact supposed to match on
arbitrary expression nodes.