Update projects to include library sources
This helps navigation. The change is done by configuring sbteclipse,
regenerating the files, and forcing Eclipse to renormalize them to
remove spurious changes and be able to check the output.

After the changes to build.sbt, the generated project file is much
closer to being usable (save for a bug on scalac options for language
options): it still contains extra directories which are not recreated on
checkout, but they are recreated when running eclipse.